+++ src/main/java/com/ccniit/graduation/dao/mongodb/IVoteDataDao.java
@@ -12,6 +12,6 @@ public interface IVoteDataDao {
 
 	List<BaseVoteData> selectVote(String tableName);
 
-	Integer counterVoteSubmitTimes(String tableName);
+	Long counterVoteSubmitTimes(String tableName);
 
 }
